payment feature

+++ cmd/web/render.go
@@ -28,7 +28,7 @@ var functions = template.FuncMap{
 	"formatCurrency": formatCurrency,
 }
 
-func formatCurrency(n float32) string {
+func formatCurrency(n int) string {
 	//f := float32(n / 100)
 	f := float32(n)
 	return fmt.Sprintf("$%.2f", f)
